added more complete database seeding models. create contacts have all firelds filled
